Reorganize rust codes
