+++ rust/src/lib_simple.rs
@@ -3,6 +3,26 @@
 
 //! Simple library entry point for Go bindings
 
-pub mod simple;
+pub mod ffi;
+pub mod runtime;
+pub mod connection;
+pub mod database;
+pub mod schema;
+pub mod table;
+pub mod metadata;
+pub mod data;
+pub mod query;
+pub mod index;
+pub mod conversion;
+pub mod types;
 
-pub use simple::*;
+// Re-export all public functions and types
+pub use ffi::*;
+pub use connection::*;
+pub use database::*;
+pub use table::*;
+pub use metadata::*;
+pub use data::*;
+pub use query::*;
+pub use index::*;
+pub use types::*;
